# web/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Koritsu API",
  "description": "Two entrances, one set of handlers: `/api` for the site (cookie session) and `/api/v1` for external clients (bearer token). Only `/api/v1` is versioned and documented.",
  "version": "2.0.0a5.4"
 },
 "paths": {
...
  "/api/jobs": {
   "post": {
    "tags": [
     "jobs"
    ],
    "summary": "Put a job into the queue",
    "description": "Accepts a job and answers at once with its card; the work happens in a worker process. Editor role in the project. A second job of the same user waits in the queue instead of being refused. 400 unknown_job_kind, 400 project_required, 402 limit_exhausted, 403 forbidden, 404 not_found, 409 in_trash.",
    "operationId": "jobs_create",
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/JobIn"
       }
      }
     }
    },
    "responses": {
     "202": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Jobs Create"
        }
       }
      }
     },
     "default": {
      "description": "Any refusal: one shape, machine-readable code",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorOut"
        }
       }
      }
     }
    }
   },
   "get": {
    "tags": [
     "jobs"
    ],
    "summary": "List your jobs",
    "description": "Your own jobs, newest first. Jobs of other people are never listed, not even in a shared project. 400 invalid_id, 400 unknown_job_status.",
    "operationId": "jobs_list",
    "parameters": [
     {
      "name": "project_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "description": "Only jobs of this project",
       "title": "Project Id"
      },
      "description": "Only jobs of this project"
     },
     {
      "name": "status",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "description": "Only jobs in this state: queued, running, done, failed, cancelled",
       "title": "Status"
      },
      "description": "Only jobs in this state: queued, running, done, failed, cancelled"
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 500,
       "minimum": 1,
       "default": 100,
       "title": "Limit"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Jobs List"
        }
       }
      }
     },
     "default": {
      "description": "Any refusal: one shape, machine-readable code",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorOut"
        }
       }
      }
     }
    }
   }
  },
  "/api/jobs/{job_id}": {
   "get": {
    "tags": [
     "jobs"
    ],
    "summary": "One job",
    "description": "One job of yours: state, progress, result or error. Asking resets the retention clock: a job is kept for as long as it is looked at. 400 invalid_id, 404 not_found.",
    "operationId": "jobs_get",
    "parameters": [
     {
      "name": "job_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Job Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Jobs Get"
        }
       }
      }
     },
     "default": {
      "description": "Any refusal: one shape, machine-readable code",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorOut"
        }
       }
      }
     }
    }
   }
  },
  "/api/jobs/{job_id}/cancel": {
   "post": {
    "tags": [
     "jobs"
    ],
    "summary": "Ask a job to stop",
    "description": "A waiting job is cancelled at once; a running one is asked to stop between steps, so whatever it has already finished is kept. 400 invalid_id, 404 not_found, 409 already_finished.",
    "operationId": "jobs_cancel",
    "parameters": [
     {
      "name": "job_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Job Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Jobs Cancel"
        }
       }
      }
     },
     "default": {
      "description": "Any refusal: one shape, machine-readable code",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorOut"
        }
       }
      }
     }
    }
   }
  },
  "/api/jobs/{job_id}/events": {
   "get": {
    "tags": [
     "jobs"
    ],
    "summary": "Events of a job since a sequence number",
    "description": "Events of one job in order, starting after `after`. Text pieces are already glued together, roughly every 100 ms. The same rows a server-sent stream reads. 400 invalid_id, 404 not_found.",
    "operationId": "jobs_events",
    "parameters": [
     {
      "name": "job_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Job Id"
      }
     },
     {
      "name": "after",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 0,
       "description": "Return events with seq greater than this",
       "default": 0,
       "title": "After"
      },
      "description": "Return events with seq greater than this"
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 2000,
       "minimum": 1,
       "default": 500,
       "title": "Limit"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Jobs Events"
        }
       }
      }
     },
     "default": {
      "description": "Any refusal: one shape, machine-readable code",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorOut"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/jobs": {
   "post": {
    "tags": [
     "v1",
     "jobs"
    ],
    "summary": "Put a job into the queue",
    "description": "Accepts a job and answers at once with its card; the work happens in a worker process. Editor role in the project. A second job of the same user waits in the queue instead of being refused. 400 unknown_job_kind, 400 project_required, 402 limit_exhausted, 403 forbidden, 404 not_found, 409 in_trash.",
    "operationId": "jobs_create_v1",
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/JobIn"
       }
      }
     }
    },
    "responses": {
     "202": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Jobs Create V1"
        }
       }
      }
     },
     "default": {
      "description": "Any refusal: one shape, machine-readable code",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorOut"
        }
       }
      }
     }
    }
   },
   "get": {
    "tags": [
     "v1",
     "jobs"
    ],
    "summary": "List your jobs",
    "description": "Your own jobs, newest first. Jobs of other people are never listed, not even in a shared project. 400 invalid_id, 400 unknown_job_status.",
    "operationId": "jobs_list_v1",
    "parameters": [
     {
      "name": "project_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "description": "Only jobs of this project",
       "title": "Project Id"
      },
      "description": "Only jobs of this project"
     },
     {
      "name": "status",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "description": "Only jobs in this state: queued, running, done, failed, cancelled",
       "title": "Status"
      },
      "description": "Only jobs in this state: queued, running, done, failed, cancelled"
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 500,
       "minimum": 1,
       "default": 100,
       "title": "Limit"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Jobs List V1"
        }
       }
      }
     },
     "default": {
      "description": "Any refusal: one shape, machine-readable code",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorOut"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/jobs/{job_id}": {
   "get": {
    "tags": [
     "v1",
     "jobs"
    ],
    "summary": "One job",
    "description": "One job of yours: state, progress, result or error. Asking resets the retention clock: a job is kept for as long as it is looked at. 400 invalid_id, 404 not_found.",
    "operationId": "jobs_get_v1",
    "parameters": [
     {
      "name": "job_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Job Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Jobs Get V1"
        }
       }
      }
     },
     "default": {
      "description": "Any refusal: one shape, machine-readable code",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorOut"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/jobs/{job_id}/cancel": {
   "post": {
    "tags": [
     "v1",
     "jobs"
    ],
    "summary": "Ask a job to stop",
    "description": "A waiting job is cancelled at once; a running one is asked to stop between steps, so whatever it has already finished is kept. 400 invalid_id, 404 not_found, 409 already_finished.",
    "operationId": "jobs_cancel_v1",
    "parameters": [
     {
      "name": "job_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Job Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Jobs Cancel V1"
        }
       }
      }
     },
     "default": {
      "description": "Any refusal: one shape, machine-readable code",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorOut"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/jobs/{job_id}/events": {
   "get": {
    "tags": [
     "v1",
     "jobs"
    ],
    "summary": "Events of a job since a sequence number",
    "description": "Events of one job in order, starting after `after`. Text pieces are already glued together, roughly every 100 ms. The same rows a server-sent stream reads. 400 invalid_id, 404 not_found.",
    "operationId": "jobs_events_v1",
    "parameters": [
     {
      "name": "job_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Job Id"
      }
     },
     {
      "name": "after",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 0,
       "description": "Return events with seq greater than this",
       "default": 0,
       "title": "After"
      },
      "description": "Return events with seq greater than this"
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 2000,
       "minimum": 1,
       "default": 500,
       "title": "Limit"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Jobs Events V1"
        }
       }
      }
     },
     "default": {
      "description": "Any refusal: one shape, machine-readable code",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorOut"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "ErrorBody": {
    "properties": {
     "code": {
      "type": "string",
      "title": "Code",
      "description": "Stable machine-readable code, e.g. not_found"
     },
     "message": {
      "type": "string",
      "title": "Message",
      "description": "Human-readable text, English"
     },
     "where": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Where",
      "description": "Field or parameter the error is about, e.g. body.email"
     }
    },
    "type": "object",
    "required": [
     "code",
     "message"
    ],
    "title": "ErrorBody",
    "description": "Тело ошибки: что случилось и, если это применимо, где."
   },
   "ErrorOut": {
    "properties": {
     "error": {
      "$ref": "#/components/schemas/ErrorBody"
     }
    },
    "type": "object",
    "required": [
     "error"
    ],
    "title": "ErrorOut",
    "description": "Единственная форма отказа службы. Другой нет ни у одного маршрута."
   },
...
   "JobIn": {
    "properties": {
     "kind": {
      "type": "string",
      "title": "Kind",
      "description": "What to do; one of: fill_tag, fill_report, agent, build, parse, export, kadai_run, kadai_rework, probe"
     },
     "project_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Project Id",
      "description": "Project the job works in; required by most kinds"
     },
     "payload": {
      "additionalProperties": true,
      "type": "object",
      "title": "Payload",
      "description": "Arguments of the job; the shape depends on the kind"
     }
    },
    "type": "object",
    "required": [
     "kind"
    ],
    "title": "JobIn",
    "description": "Тело постановки задания.\n\nИмя по-английски, как у соседей (`ProjectPatchIn`, `ModelKeyIn`): оно уезжает\nв OpenAPI и становится именем типа в клиенте сайта."
   },
...
  }
 },
...
}